+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 ###### (auto syncing with the upstream repository)
 
+## 20240205
+* luci-app-dockerman✖️
+* crun✔
+* podman✔
+* netavark: bomp version to 1.10.2 for native support nftables.
+
 ---
 
 ## 20240129